Fix bug with single-syllable words
While typing around in the webpage, I noticed that sometimes, for short
words, no part of the word would be printed bold. This turns out to
happen when the word has a single syllable and begins with a consonant
(“cair”, as in “Cair Andros”, is a real example for this); from the
point of view of the analyse() algorithm, this syllable “closes” itself
(viewed from the end – unlike “alph”, which remains “open”) and doesn’t
use the “push the syllable break at the beginning” case, which led its
stressed syllable to be off by one (-1 rather than 0). Fix this by
moving the “no naturally stressed syllable” case *after* the syllable
break at the beginning is pushed, and always adjusting for the
off-by-one case. This fixes “cair” without breaking any of the other
test cases.