Respect prefers-reduced-motion preference
Updating the output on every keystroke may be a bit much for some users;
if they’ve instructed their user agent that they prefer reduced motion,
only update the word when they’ve “committed” a change to the input
(e. g. by pressing Enter or tabbing out of it). And since in this case
it might not be completely clear what you have to do to update the
output, also add a button to do it explicitly.