Fix Sindarin au diphthong
The function mistakenly believed that au was only a diphthong in Quenya,
which led to the claim that Naugrim and Nauglamír (Sindarin words: the
Quenya is Nauco) began na|UG – absurd.

This also gets rid of a fall-through – instead, let’s just set the vowel
flag at the top, and then undo it in the one case where it shouldn’t be
set (qu = cw).