Avoid one toUpperCase() in tests
The input syllables are all either all-lowercase or all-uppercase, so to
determine if a syllable is all-uppercase or not, we can just check
whether it’s equal to its all-lowercased version, which we already
needed anyways, and avoid a separate toUpperCase() call.

Also, declare the iteration variables as const instead of let, just
because we can.