Add explanation about “muta cum liquida” rule
There is some evidence that it should be IM|la|dris, NAR|go|thrond, and
by extension maybe ME|ne|groth also, and other words may also be meant
to be stressed differently. This has already been puzzled out by others,
so we can link to someone else’s better essay. Our “get out of jail free
card” is that the rule does not appear to be systematic, which is the
excuse to avoid updating the tool for it. (I suppose I could make the
tool show both possibilities in these cases, but for now I can’t be
bothered to do that.)