Fix analysis of “Orophin” (PH)
It seems like, in the middle of a word, PH always represents -ff-, and
so counts as two consonants. I searched Wikidata for Tolkien’s universe
labels containing PH, and the relevant examples (Oropher, Araphant,
Celepharn, Araphor, Orophin) are all claimed by The Encyclopedia of Arda
[1] to be stressed on the second-to-last syllable. Orophin is the name
of a Silvan Elf, so for the test case I assume it’s Sindarin.

[1]: https://www.glyphweb.com/arda/o/oropher.php